Rename lib to utils
